+++ CoreRemoting/RemotingClient.cs
@@ -19,6 +19,7 @@ using CoreRemoting.Serialization.Bson;
 using CoreRemoting.Toolbox;
 using CancellationTokenSource = System.Threading.CancellationTokenSource;
 using Timer = System.Timers.Timer;
+using CoreRemoting.Threading;
 
 namespace CoreRemoting
 {
